Separting in files components
